services/auth_service/requirements.txt
Django==5.0.7
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==23.5
drf-yasg==1.21.7
Pillow==10.2.0
django-storages==1.14.2
boto3==1.34.69 